% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_codings.R
\name{extract_codings}
\alias{extract_codings}
\title{Extract coded text fragments from MER file}
\usage{
extract_codings(fname)
}
\arguments{
\item{fname}{path to the MER file}
}
\value{
Data frame with columns
\itemize{
\item \code{ID} - ID of the fragment
\item \code{TextID} - ID of the source document
}
}
\description{
Extract all coded text fragments from MAXQDA exchange file included, among
other things, start and end character positions of each encoded text
fragment.
}
