% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_codings.R
\name{extract_mex_codings}
\alias{extract_mex_codings}
\title{Extract coded text fragments from MEX file}
\usage{
extract_mex_codings(fname)
}
\arguments{
\item{fname}{path to the MEX file}
}
\value{
Data frame with columns:
\itemize{
\item \code{ID} - ID of the fragment
\item \code{TextID} - ID of the source document
\item \code{WordID} - ID of the codeword
\item \code{Preview} - First 63 characters of the coded text fragment
\item \code{start} - Position of the initial character of the coded text fragment
\item \code{end} - Position of the final character of the coded text fragment
\item \code{file_name} - Name of the file with the source text
\item \code{tag} - Name of the codeword
\item \code{preview_length} - Number of characters in the Preview string
}
}
\description{
Extract all coded text fragments from MAXQDA exchange file included, among
other things, start and end character positions of each encoded text
fragment.
}
